Move SavingAnimation and SaveNoteEditor to new save-page-message out of the old drawer folder
