+++ src/chrome-extension/content-script/components/save-page-message/components/page-saving-animation/saving-animation/saving-animation.scss
@@ -1,4 +1,4 @@
-@import "../../../../../constants/variables";
+@import "../../../../../../../constants/variables";
 .submitting-container {
   display: flex;
   flex-direction: column;
