+++ src/chrome-extension/content-script/components/save-page-message/components/save-page-content/save-note-editor/SaveNoteEditor.tsx
@@ -4,9 +4,9 @@ import {Slate, ReactEditor} from "slate-react";
 import {EditablePlugins, pipe} from "@udecode/slate-plugins";
 import {createEditor, Node} from "slate";
 import {equals} from "ramda";
-import {baseKeyBindingHandler} from "../../../../../common/rich-text-editor/utils/keyboard-handler";
-import { chromeExtensionPlugins } from "../../../../../common/rich-text-editor/editors/chrome-extension/config";
-import {CHROME_EXTENSION} from "../../../../../common/rich-text-editor/editors/chrome-extension/constants";
+import {baseKeyBindingHandler} from "../../../../../../../common/rich-text-editor/utils/keyboard-handler";
+import { chromeExtensionPlugins } from "../../../../../../../common/rich-text-editor/editors/chrome-extension/config";
+import {CHROME_EXTENSION} from "../../../../../../../common/rich-text-editor/editors/chrome-extension/constants";
 
 export const SaveNoteEditor = (props: { content: Node[], setContent: (newValue: Node[]) => void, editor: ReactEditor }) => {
     const { content, setContent, editor } = props
